+++ apps/test-app/app/tests/toolbar/index.tsx
@@ -3,7 +3,8 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-import { IconButton } from "@stratakit/bricks";
+import * as React from "react";
+import { Divider, IconButton } from "@stratakit/bricks";
 import { unstable_Toolbar as Toolbar } from "@stratakit/structures";
 import { definePage } from "~/~utils.tsx";
 
@@ -11,6 +12,10 @@ import placeholderIcon from "@stratakit/icons/placeholder.svg";
 
 export default definePage(
 	function Page() {
+		const [active, setActive_] = React.useState("");
+		const setActive = (id: string) => {
+			setActive_((prev) => (prev === id ? "" : id));
+		};
 		return (
 			<Toolbar.Group variant="solid">
 				<Toolbar.Item
@@ -19,6 +24,8 @@ export default definePage(
 							icon={`${placeholderIcon}#icon-large`}
 							label="Click me"
 							variant="ghost"
+							isActive={active === "1"}
+							onClick={() => setActive("1")}
 						/>
 					}
 				/>
@@ -28,6 +35,8 @@ export default definePage(
 							icon={`${placeholderIcon}#icon-large`}
 							label="Click me"
 							variant="ghost"
+							isActive={active === "2"}
+							onClick={() => setActive("2")}
 						/>
 					}
 				/>
@@ -37,18 +46,27 @@ export default definePage(
 							icon={`${placeholderIcon}#icon-large`}
 							label="Click me"
 							variant="ghost"
+							isActive={active === "3"}
+							onClick={() => setActive("3")}
 						/>
 					}
 				/>
 			</Toolbar.Group>
 		);
 	},
-	{ visual: VisualTest },
+	{
+		visual: VisualTest,
+		vertical: () => <TestToolbar orientation="vertical" />,
+	},
 );
 
-function VisualTest() {
+interface TestToolbarProps {
+	orientation: "horizontal" | "vertical";
+}
+
+function TestToolbar({ orientation }: TestToolbarProps) {
 	return (
-		<Toolbar.Group variant="solid">
+		<Toolbar.Group variant="solid" orientation={orientation}>
 			<Toolbar.Item
 				render={
 					<IconButton
@@ -58,12 +76,16 @@ function VisualTest() {
 					/>
 				}
 			/>
+			<Divider
+				orientation={orientation === "horizontal" ? "vertical" : "horizontal"}
+			/>
 			<Toolbar.Item
 				render={
 					<IconButton
 						icon={`${placeholderIcon}#icon-large`}
 						label="Click me"
 						variant="ghost"
+						isActive
 					/>
 				}
 			/>
@@ -77,5 +99,21 @@ function VisualTest() {
 				}
 			/>
 		</Toolbar.Group>
+	);
+}
+
+function VisualTest() {
+	return (
+		<div
+			style={{
+				display: "flex",
+				gap: "var(--stratakit-space-x2)",
+				flexDirection: "column",
+				alignItems: "flex-start",
+			}}
+		>
+			<TestToolbar orientation="horizontal" />
+			<TestToolbar orientation="vertical" />
+		</div>
 	);
 }