+++ packages/structures/CHANGELOG.md
@@ -1,5 +1,66 @@
 # Changelog
 
+## 0.4.0
+
+### Breaking changes
+
+- [#888](https://github.com/iTwin/design-system/pull/888): `Toolbar.Item` component no longer automatically uses the large version of the icon.
+
+  `#icon-large` must now be explicitly added to the URL to select the large icons from `@stratakit/icons`. For example:
+
+  ```diff
+    <Toolbar.Item
+  -   render={<IconButton icon={placeholderIcon} />}
+  +   render={<IconButton icon={`${placeholderIcon}#icon-large`} />}
+    />
+  ```
+
+- [#900](https://github.com/iTwin/design-system/pull/900): Renamed `Tabs.Root` component to `Tabs.Provider`.
+
+  ```diff
+  - <Tabs.Root>
+  + <Tabs.Provider>
+      <Tabs.TabList>…</Tabs.TabList>
+      <Tabs.TabPanel>…</Tabs.TabPanel>
+  - </Tabs.Root>
+  + </Tabs.Provider>
+  ```
+
+  This change makes StrataKit's naming convention more consistent. `Root` components always render a DOM element whereas `Provider` components have no underlying DOM element.
+
+- [#900](https://github.com/iTwin/design-system/pull/900): Renamed `DropdownMenu.Root` component to `DropdownMenu.Provider`.
+
+  ```diff
+  - <DropdownMenu.Root>
+  + <DropdownMenu.Provider>
+      <DropdownMenu.Button>…</DropdownMenu.Button>
+      <DropdownMenu.Content>…</DropdownMenu.Content>
+  - </DropdownMenu.Root>
+  + </DropdownMenu.Provider>
+  ```
+
+  This change makes StrataKit's naming convention more consistent. `Root` components always render a DOM element whereas `Provider` components have no underlying DOM element.
+
+### Non-breaking changes
+
+- [#903](https://github.com/iTwin/design-system/pull/903): Added proper styling for `Divider` rendered inside a `Toolbar.Group`.
+
+  ```tsx
+  <Toolbar.Group variant="solid">
+    <Toolbar.Item render={…} />
+    <Divider orientation="vertical" />
+    <Toolbar.Item render={…} />
+    <Toolbar.Item render={…} />
+  </Toolbar.Group>
+  ```
+
+- [#913](https://github.com/iTwin/design-system/pull/913): Updated internal CSS selectors in every component.
+- [#901](https://github.com/iTwin/design-system/pull/901): Added `orientation` prop to `Toolbar.Group` component. Set the `orientation` prop to `"vertical"` to display the toolbar vertically.
+- [#902](https://github.com/iTwin/design-system/pull/902): Updated active style of a ghost `IconButton` when used in a `Toolbar` component.
+- Updated dependencies:
+  - @stratakit/foundations@0.3.0
+  - @stratakit/bricks@0.4.0
+
 ## 0.3.2
 
 - [#881](https://github.com/iTwin/design-system/pull/881): Updated CSS to use `--stratakit-space-` variables instead of hardcoded values in all components.